requirements.txt
requests
python-dotenv
autobahn
twisted
